requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langgraph>=0.2.0
anthropic>=0.40.0
google-genai>=1.0.0
openai>=1.50.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
slowapi>=0.1.9
